requirements.txt
python-dateutil>=1.5,<2.5
requests-oauthlib>=0.6.1,<1.1
httplib2>=0.9.2
google_api_python_client>=1.5.1
parsedatetime>=2.1
oauth2client>=3.0,<4.0
